requirements.txt
asgiref==3.3.1; python_version >= "3.7"
django==3.1.7; python_version >= "3.6"
djangorestframework-simplejwt==4.6.0; python_version >= "3.7"
djangorestframework==3.12.3; python_version >= "3.7"
psycopg2-binary==2.8.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pyjwt==2.0.1; python_version >= "3.7"
pytz==2021.1; python_version >= "3.7"
sqlparse==0.4.1; python_version >= "3.7"
